\name{plot.cfit}
\alias{plot.cfit}
\title{Plot Command For Class Cfit}
\description{
  Plot Command For Class Cfit
}
\usage{
plot.cfit <- function(fit)
}
\arguments{
  \item{fit}{
    object of class \code{cfit}
  }
}
\value{
  a plot
}
\seealso{
  \code{\link{rho}}, \code{\link{pion}}
}
\author{Carsten Urbach, \email{carsten.urbach@liverpool.ac.uk}}
\keyword{methods}
\keyword{hplot}

